refactor hdl_specs Signed-off-by: chipmuenk <mail@chipmuenk.de>
